+++ index.html
@@ -152,14 +152,15 @@
           <!-- Комментарии к изображению -->
           <div class="social__comment-count">5 из <span class="comments-count">125</span> комментариев</div>
           <ul class="social__comments">
-            <li class="social__comment">
+
+<!--             <li class="social__comment">
               <img class="social__picture" src="img/avatar-4.svg" alt="Аватар комментатора фотографии" width="35" height="35">
               <p class="social__text">Мега фото! Просто обалдеть. Как вам так удалось?</p>
             </li>
             <li class="social__comment">
               <img class="social__picture" src="img/avatar-3.svg" alt="Аватар комментатора фотографии" width="35" height="35">
                <p class="social__text">Да это фоташоп!!!!!!!!</p>
-            </li>
+            </li> -->
           </ul>
 
           <!-- Кнопка для загрузки новой порции комментариев -->
@@ -203,6 +204,14 @@
         <span class="picture__likes"></span>
       </p>
     </a>
+  </template>
+
+    <!-- Шаблон коментария случайного пользователя -->
+  <template id="comment">
+    <li class="social__comment">
+      <img class="social__picture" src="" alt="Аватар комментатора фотографии" width="35" height="35">
+      <p class="social__text"></p>
+    </li>
   </template>
 
   <!-- Сообщение с ошибкой загрузки изображения -->
